pyproject.toml
[project]
name = "finance-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "dany1940",email = "59851426+dany1940@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "confluent-kafka (>=2.8.0,<3.0.0)",
    "pyspark (>=3.5.4,<4.0.0)",
    "polars (>=1.21.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "yfinance (>=0.2.52,<0.3.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"

